Move Grammer Logic to New Parser Logic
I realized that I was moving towards a general-purpose parsing tool when
I starting thinking ASTs. I wrote a generic parser that I can define a
subset of SQL syntax with. Whether I'm parsing SQL as a text or a
sequence of tokens from the builders, I should be able to analyze the
output and generate the corresponding output. Things are getting
interesting...