Support Quantifiers
I've always debated supporting EXISTS. I finally decided to add support,
if not just to handle SQL when converting to command builders. Many
developers use EXISTS over IN because of performance reasons. I'm sure
they'd be annoyed if it was off limits.

I also now support the other quantifiers (ALL, ANY and SOME). These are
rare filters (I'd never heard of them) but are considered part of
standard SQL. Technically these could be implemented as function
calls... but I want to avoid wacky workarounds like that and directly
support the feature.

I created a Literal base class. I may use this in the future to simplify
the creation of literals.

All select combiners (UNION, INTERSECT, EXCEPT and MINUS) now accept a
DISTINCT qualifier (ALL or DISTINCT). This eliminates the need for the
UnionAll class.