Default to placeholders in some cases
In cases where it is unclear what an identifier is, don't assume it is a
column name. Instead, only make it a column if it is qualified with the
source name or if there is a single source. Additionally, check for a
prefix to treat an identifier as a placeholder (parameter) rather than a
column.